Let the debugging IDE listen on port 9000
The way xdebug works, the debugging program listens on port 9000
and the xdebug service connects to that program over that port.
So we have to prevent docker from binding to the port so the
debugging program (e.g., PHPStorm) can listen on it. Removed
the mapping for port 9000 from docker/docker-compose.yml.

Also added .vscode to the .gitignore file.