Persist database via a named volume. (Possible fix for #15)
The Docker Library project documents a problem with some versions of MariaDB
not starting correctly when a host system directory is mounted to /var/lib/mysql.
https://github.com/docker-library/mariadb/issues/38

It appears to work correctly when a named volume is used; at least,
it works on Windows.  This should be tested on OS/X as well
before being merged.