Fix Drupal moderation handling of new translations
Applies the patch for https://www.drupal.org/project/drupal/issues/3029219.
That bug did not correctly recognize that a new translation was
at the beginning of its moderation state cycle, instead conflating
the moderation history of the original language content with that
of the new translation. This resulted in failed saves because of
rejection of what should have been accepted as valid state transitions.

The Drupal core team seems reluctant to accept our patch (or even
acknowledge that this is a bug), so we have no choice but to move on
and just apply the patch ourselves.

Closes #475