Removing memcache_enabled check.
Removed at the suggestion of our TAM, "The flag is going to be removed in BLT soon. After the updates are finished to allow us to use the new version of memcache we're going to pull the flags out. Right now they rely on a setting inside the ACSF dashboard which is only available to Acquians."