(#3275) Fix sprite paths.
This is a short term fix as all of our image references are pretty
questionable. There is a mix of absolute references and some references using
Aliases. Alias end up resolving to relative paths.

The initial issue was that the previous fix to update the loaders
bumped css-loader, which now looks at absolute paths. So a fix was
added to ignore all urls from resolution. This is where things went
wrong. This mean Aliases were no longer resolved.

So TL;DR; we:
- added a new Alias for the Dist/Image folder for sprites.
- updated calls to sprites to reference that alias, which follows the
  previous file-loader outputPath
- Updated the css-loader config to ignore absolute urls. This probably
  should be fixed at some point to use the ImageDist alias.