(#3513) Add CTHP card title field
This will allow the field_browser_title value, populated from
the CDR, to be used as for the browser title, as was originally
intended.

Tasks include:
- create new field_ctrp_card_title field, 100 characters max
- modify the RESTful services to handle the new field
- modify the tests to include the new field
- adjust the YAML sample content to include the new field
- add the new field to the CTHP: Guide Card List view
- modify the cthp guide card list item template to use new field
- use the field_browser_title value for the head/title element

Deployment will be coupled with a full load of the CIS summaries,
to populate the new field with appropriate values, and to adjust
the values in the browser title field so that they are unique.